code cleanup for crypto modules.
