+++ mods/saitocrypto/saitocrypto.js
@@ -4,35 +4,38 @@ const { Keyring }  = require('@polkadot/keyring');
 const { mnemonicGenerate } = require('@polkadot/util-crypto');
 const { ApiPromise, WsProvider }  = require('@polkadot/api');
 const { randomBytes }   = require('crypto');
-class SaitoCrypto extends ModTemplate {
 
+// Stub module for future integration if we want Saito to do is_cryptocurrency
+class SaitoCrypto extends ModTemplate {
   constructor(app) {
 
     super(app);
-
     this.name = "SaitoCrypto";
-    this.description = "Payment Gateway for DOT cryptocurrency in Saito Application";
+    this.description = "Saito";
     this.categories = "Crptocurrency";
     this.ticker = "SAITO";
     this.mods = [];
+    this.app = app;
   }
   
   requestInterface(type = "") {
     if (type == "is_cryptocurrency") {
       return {
-        // The ticker of the cryptocurrency
+        name: this.name,
+        description: this.description,
         ticker: this.ticker,
-        // get balance of given address.
-        //  getBalance(address)
-        getBalance: async() => { return 0; },
-        // transfer to another addresses
-        transfer: null,
-        // sign and send a raw transaction
-        signAndSend: null,
+        getBalance: async() => {
+          return this.app.wallet.returnBalance();
+        },
+        transfer: async(howMuch, to) => {
+          console.log("implement me!!");
+        },
+        getAddress: async() => {
+          return this.app.wallet.returnPublicKey();
+        },
         subscribe: null,
-        getAddress: async() => { return "asdf"; },
-        getKeyring: null,
-        buildRawTx: null,
+        //https://polkadot.js.org/docs/api/cookbook/tx/
+        estimateFee: null,
       }
     }
     return null;
@@ -40,9 +43,6 @@ class SaitoCrypto extends ModTemplate {
   initialize(app) {
     super.initialize(app);
   }
-
-    
-
 }
 
 module.exports = SaitoCrypto;
